+++ load.h
@@ -5,8 +5,8 @@ void convertcol(long col, struct Polygon *poly);
 /* arrays are allocated during the load since the number of vertices per */
 /* polygon can vary greatly. */
 
-short loadobject(char *file);
+int loadobject(char *file);
 
 /* Load viewing options and store in the global 'vopts' structure. */
 
-short loadvopts(char *file);
+int loadvopts(char *file);
